Added option to remove db in local. Remote is still nor supported.

+++ template-scripts/benchmark-mongodb-a.py
@@ -1,6 +1,7 @@
 # Script for execute workload A on Mongodb with various parameters.
 import subprocess
 import plots
+import sys
 
 REPO_HOME="=REPO_DIR="
 HOME_YCSB="=REPO_DIR=/ycsb-0.17.0"
@@ -26,10 +27,15 @@ def execute(command) :
     process = subprocess.call(command, shell=True)
     return process
 
-def delete_db() :
+def delete_db(usage) :
     print("\n\n==> Removing db")
-    command = "docker exec -it mongo mongo ycsb --eval \\\"db.dropDatabase()\\\""
-    execute(command)
+    if usage=="local":
+        command = "docker exec -it mongo mongo ycsb --eval \\\"db.dropDatabase()\\\""
+        execute(command)
+    else:
+        print("Please, manually delete the db. This function is still not supported.")
+        input()
+    
 
 def laod_data(dim) :
     # Execute workload a for every dim
@@ -41,6 +47,14 @@ def run_workload(dim, t) :
     command = BIN_YCSB + "/ycsb.sh run mongodb-async -s -P " + HOME_YCSB + "/workloads/workloada -p recordcount=" + str(dim) + " -target " + str(t) + " -p mongodb.url=" + mongodb_connection_string + " > " + getOutputFilename("Run", dim, t)
     execute(command)
 
+
+if len(sys.argv)<2 :
+    print("Error, required parameter: ")
+    print("- Usage: local or remote")
+    print("Usage: py benchmark-mongodb-a.py <usage>")
+    exit(1)
+
+
 # Every record is 1KB (10 fields of 100B)
 input_dim = [1000, 10000, 100000, 500000, 1000000]
 throughput = [100, 1000, 100000]
@@ -50,10 +64,11 @@ for dim in input_dim:
     laod_data(dim)
     for t in throughput:
         print("\n==> Wordload A. Input dim: " + str(dim) + "; Throughput: " + str(t))
-        laod_data(dim)
         run_workload(dim, t)
     
-    delete_db()
+    delete_db(sys.argv[1])
+
+# PROSSIMI PASSI: AGGIUNGERE ITERAZIONE SUI CLIENT.
 
 plots.generateAndShowInputLatencyPlots(input_dim, throughput, "test-mongo-scenarioA")
 plots.generateAndShowThroughputLatencyPlots(input_dim, throughput, "test-mongo-scenarioA")
